FutureWarning and DeprecationWarning tidy
Tidied these sorts of warnings in the code:
pyTA.py:1447: FutureWarning: The squeeze argument has been deprecated and will be removed in a future version. Append .squeeze("columns") to the call to squeeze.
pyTA.py:152: DeprecationWarning: an integer is required (got type numpy.float64).  Implicit conversion to integers using __int__ is deprecated, and may be removed in a future version of Python.