Confirmed motor cooscillation works
Works well and can take measurements with the motors running!

Key next step motors-wise is to try and get it working with one motor only.

Small issue: 'Hardware' tab gets re-activated when you stop either a measurement (acq, dia) or the motors, so one or the other is still running. Need to prevent reactivation so that it's not possible to disconnect hardware while while it's still being used.